requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
pymongo==4.5.0
